Chore : Create Extension group and add file
